refactor: ♻️ encapsulate project detail dpp
