+++ pages/project/[id]/index.tsx
@@ -24,11 +24,9 @@ import { useTranslation } from "next-i18next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { useRouter } from "next/router";
 import { useCallback, useEffect, useRef, useState } from "react";
-import Tree from "react-d3-tree";
 
 // Components
-import { Button, Card, Frame, Icon, Modal, Spinner, Stack, Tabs, Text, Toast } from "@bbtgnn/polaris-interfacer";
-import { DuplicateMinor, MaximizeMinor } from "@shopify/polaris-icons";
+import { Button, Spinner, Stack, Tabs, Text } from "@bbtgnn/polaris-interfacer";
 import BrBreadcrumb from "components/brickroom/BrBreadcrumb";
 import FullWidthBanner from "components/FullWidthBanner";
 import ProjectDetailOverview from "components/ProjectDetailOverview";
@@ -43,6 +41,7 @@ import { Cube, Events, ListBoxes, ParentChild, Purchase } from "@carbon/icons-re
 import BrThumbinailsGallery from "components/brickroom/BrThumbinailsGallery";
 import ContributionsTable from "components/ContributionsTable";
 import ContributorsTable from "components/ContributorsTable";
+import ProjectDpp from "components/partials/project/ProjectDpp";
 import ProjectSidebar from "components/partials/project/[id]/ProjectSidebar";
 import ProjectTypeChip from "components/ProjectTypeChip";
 
@@ -104,51 +103,10 @@ const Project = () => {
     setInList(_listParsed.includes(project?.id));
   }, [project, getItem]);
 
-  const handleCollect = () => {
-    const _list = getItem("projectsCollected");
-    const _listParsed = _list ? JSON.parse(_list) : [];
-    if (_listParsed.includes(project!.id)) {
-      setItem("projectsCollected", JSON.stringify(_listParsed.filter((a: string) => a !== project!.id)));
-      setInList(false);
-    } else {
-      const _listParsedUpdated = [..._listParsed, project?.id];
-      setItem("projectsCollected", JSON.stringify(_listParsedUpdated));
-      setInList(true);
-    }
-  };
-
   // Tabs setup
   const handleTabChange = useCallback((selectedTabIndex: number) => setSelected(selectedTabIndex), []);
 
   //
-
-  const [active, setActive] = useState(false);
-  const toggleActive = useCallback(() => setActive(active => !active), []);
-
-  function copyDPP() {
-    navigator.clipboard.writeText(JSON.stringify(data?.economicResource.traceDpp, null, 2));
-    setActive(true);
-  }
-
-  // map trace dpp to treeData
-  const dppToTreeData = (dpp: any) => {
-    return dpp?.children?.map((child: any) => {
-      return {
-        name: child.type,
-        children: dppToTreeData(child),
-        attributes: { name: child.node.name || child.node.action_id },
-      };
-    });
-  };
-
-  const [activeTree, setActiveTree] = useState(false);
-
-  // DPP Tree
-  const translate = { x: width / 2, y: 20 };
-  const treeData = dppToTreeData(data?.economicResource.traceDpp);
-
-  //
-
   // Check if project has been just created
   const isCreated = router.query.created === "true";
 
@@ -286,80 +244,7 @@ const Project = () => {
 
               {selected == 0 && <ProjectDetailOverview project={project} />}
               {selected == 1 && <RelationshipTree project={data?.economicResource} />}
-              {selected == 2 && (
-                <div className="space-y-8">
-                  <div>
-                    <div className="flex justify-between items-center mb-4">
-                      <Text as="h1" variant="headingXl">
-                        {t("Tree view")}
-                      </Text>
-                      <div className="space-x-2">
-                        <Button onClick={copyDPP} icon={<Icon source={DuplicateMinor} />}>
-                          {t("Copy DPP")}
-                        </Button>
-                      </div>
-                    </div>
-                    <div className="border-1 border-border-subdued bg-white rounded-md h-64 relative">
-                      <div className="absolute bottom-4 right-4">
-                        <Button onClick={() => setActiveTree(true)} icon={<Icon source={MaximizeMinor} />}>
-                          {t("Full screen")}
-                        </Button>
-                      </div>
-                      <Tree
-                        data={treeData}
-                        orientation="vertical"
-                        nodeSize={{ x: 300, y: 100 }}
-                        translate={{ x: 150, y: 20 }}
-                        zoom={0.5}
-                        hasInteractiveNodes={false}
-                        zoomable={false}
-                      />
-                    </div>
-                  </div>
-
-                  <div>
-                    <div className="flex justify-between items-center mb-4">
-                      <Text as="h1" variant="headingXl">
-                        {t("JSON view")}
-                      </Text>
-                      <Button onClick={copyDPP} icon={<Icon source={DuplicateMinor} />}>
-                        {t("Copy DPP")}
-                      </Button>
-                    </div>
-                    <DynamicReactJson
-                      src={data?.economicResource.traceDpp}
-                      collapsed={3}
-                      enableClipboard={true}
-                      displayDataTypes={false}
-                      sortKeys={true}
-                    />
-                  </div>
-                </div>
-              )}
-              <Modal
-                fullScreen
-                large
-                open={activeTree}
-                onClose={() => setActiveTree(false)}
-                title={t("Digital Product Passport Tree")}
-              >
-                <Modal.Section>
-                  <div className="flex justify-end">
-                    <Text as="p" variant="bodyMd" color="subdued">
-                      {t("Pan to move, pinch to zoom")}
-                    </Text>
-                  </div>
-                  <div className="h-[100vh]" ref={ref}>
-                    <Tree
-                      data={treeData}
-                      zoom={1}
-                      translate={{ x: 200, y: 50 }}
-                      orientation="vertical"
-                      nodeSize={{ x: 300, y: 100 }}
-                    />
-                  </div>
-                </Modal.Section>
-              </Modal>
+              {selected == 2 && <ProjectDpp id={data?.economicResource.id!} />}
 
               {selected == 3 && (
                 <ContributorsTable
@@ -377,8 +262,6 @@ const Project = () => {
         <div className="hidden lg:block">{sidebar}</div>
         {/* Sidebar */}
       </div>
-
-      <Frame>{active ? <Toast content={t("DPP copied!")} onDismiss={toggleActive} duration={2000} /> : null}</Frame>
     </>
   );
 };
